Restructure how pages are being imported
