+++ src/pages/404/NotFound.scss
@@ -1,6 +1,6 @@
 .container {
   height: 100vh;
-  background: #ffffff url("../assets/images/background.png") no-repeat center;
+  background: #ffffff url("../../assets/images/background.png") no-repeat center;
   background-size: contain;
 }
