+++ src/pages/index.js
@@ -1,3 +1,4 @@
-import NotFound from "./NotFound";
+import NotFound from "./404/NotFound";
+import CounterPage from "./Counter/index"
 
-export default NotFound;
+export  { NotFound, CounterPage };
